+++ packages/scripts/src/components/front-end/plugin/freeform.ts
@@ -7,6 +7,7 @@ import DragAndDropHandler from "@lib/plugin/handlers/fields/drag-and-drop";
 import InputMaskHandler from "@lib/plugin/handlers/fields/input-mask";
 import RatingHandler from "@lib/plugin/handlers/fields/rating";
 import SignatureHandler from "@lib/plugin/handlers/fields/signature";
+import SummaryHandler from "@lib/plugin/handlers/fields/summary";
 import TableHandler from "@lib/plugin/handlers/fields/table";
 import AbTestHandler from "@lib/plugin/handlers/form/ab-test";
 import GoogleTagManager from "@lib/plugin/handlers/form/google-tag-manager";
@@ -89,6 +90,7 @@ export default class Freeform {
     DragAndDropHandler,
     SaveFormHandler,
     CardsHandler,
+    SummaryHandler,
   ];
 
   _lastButtonPressed?: HTMLButtonElement;